Move ArrayAccessObject to test objects.
