+++ tests/ArrayAccessTraitTest.php
@@ -6,6 +6,7 @@ namespace Yiisoft\Arrays\Tests;
 
 use ArrayIterator;
 use PHPUnit\Framework\TestCase;
+use Yiisoft\Arrays\Tests\Objects\ArrayAccessObject;
 
 class ArrayAccessTraitTest extends TestCase
 {
